logout menu item appears only when a user has logged into the app.
